requirements.txt
alembic==0.8.8
aniso8601==1.1.0
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
click==6.6
coverage==4.2
decorator==4.0.10
enum34==1.1.6
Flask==0.11.1
Flask-Migrate==2.0.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
nose==1.3.7
passlib==1.6.5
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.7
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
